Did Tons of Tests on Delete and Edit
I didn't use the features of JUnit tester very much because I found it a lot better to have all of the files in front of me and go through them. Did a bunch of test cases with delete and edit to make sure everything worked. Now just going to write some comments so the next person won't suffer for my weird commit code.